Backport (read: copied) net/wireless and drivers/net/wireless from 3.…
…0.8 for bcmdhd

These are the version as of e0de0a5
on common/android-3.0.

The bcmdhd module allows Wi-Fi direct to be used for bcm4329 and
bcm4330. The module is now compiled into the kernel (the same
as Galaxy Nexus) as the system won't load the module correctly
for Wi-Fi direct and it sometimes crashes the system.

Both the wifi and tethering interfaces are renamed wlan0 with this
patch.